e2e-tests/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via requests
colorama==0.4.6
    # via pytest
coverage[toml]==7.2.5
    # via pytest-cov
exceptiongroup==1.1.1
    # via pytest
flake8==6.0.0
    # via -r requirements.in
greenlet==2.0.1
    # via playwright
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
mccabe==0.7.0
    # via flake8
mypy==1.3.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via mypy
packaging==23.1
    # via pytest
playwright==1.33.0
    # via pytest-playwright
pluggy==1.0.0
    # via pytest
pycodestyle==2.10.0
    # via flake8
pyee==9.0.4
    # via playwright
pyflakes==3.0.1
    # via flake8
pytest==7.3.1
    # via
    #   -r requirements.in
    #   pytest-base-url
    #   pytest-cov
    #   pytest-dotenv
    #   pytest-playwright
pytest-base-url==2.0.0
    # via pytest-playwright
pytest-cov==4.0.0
    # via -r requirements.in
pytest-dotenv==0.5.2
    # via -r requirements.in
pytest-playwright==0.3.3
    # via -r requirements.in
python-dotenv==1.0.0
    # via pytest-dotenv
python-slugify==8.0.1
    # via
    #   -r requirements.in
    #   pytest-playwright
requests==2.30.0
    # via pytest-base-url
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   mypy
    #   pytest
types-python-slugify==8.0.0.2
    # via -r requirements.in
typing-extensions==4.5.0
    # via
    #   mypy
    #   pyee
urllib3==2.0.2
    # via requests
